assignment-2-103704559/code/mysimbdp/requirements.platform.txt
fastapi
uvicorn
docker
confluent-kafka
cassandra-driver
requests
pyyaml
apscheduler
google-cloud-storage